Change i32 literal to f32

+++ src/operations.rs
@@ -9,7 +9,7 @@ pub trait LpOperations<T> where T: Into<LpExpression> {
     fn equal(&self, lhs_expr: T) -> LpConstraint;
 }
 
-impl Into<LpExpression> for i32 {
+impl Into<LpExpression> for f32 {
     fn into(self) -> LpExpression {
         LitVal(self)
     }
@@ -36,7 +36,7 @@ impl<T: Into<LpExpression> + Clone, U> LpOperations<T> for U where U: Into<LpExp
 }
 
 
-// LpExpr + (LpExpr, &LpExpr, i32)
+// LpExpr + (LpExpr, &LpExpr, f32)
 impl<T> Add<T> for LpExpression where T: Into<LpExpression> + Clone {
     type Output = LpExpression;
     fn add(self, _rhs: T) -> LpExpression {
@@ -44,7 +44,7 @@ impl<T> Add<T> for LpExpression where T: Into<LpExpression> + Clone {
     }
 }
 
-// &LpExpr + (LpExpr, &LpExpr, i32)
+// &LpExpr + (LpExpr, &LpExpr, f32)
 impl<'a, T> Add<T> for &'a LpExpression where T: Into<LpExpression> + Clone {
     type Output = LpExpression;
     fn add(self, _rhs: T) -> LpExpression {
@@ -52,15 +52,15 @@ impl<'a, T> Add<T> for &'a LpExpression where T: Into<LpExpression> + Clone {
     }
 }
 
-// i32 + &LpExpr
-impl<'a> Add<&'a LpExpression> for i32 {
+// f32 + &LpExpr
+impl<'a> Add<&'a LpExpression> for f32 {
     type Output = LpExpression;
     fn add(self, _rhs: &'a LpExpression) -> LpExpression {
         AddExpr(Rc::new(LitVal(self)), Rc::new(_rhs.clone()))
     }
 }
 
-// LpExpr - (LpExpr, &LpExpr, i32)
+// LpExpr - (LpExpr, &LpExpr, f32)
 impl<T> Sub<T> for LpExpression where T: Into<LpExpression> + Clone {
     type Output = LpExpression;
     fn sub(self, _rhs: T) -> LpExpression {
@@ -68,7 +68,7 @@ impl<T> Sub<T> for LpExpression where T: Into<LpExpression> + Clone {
     }
 }
 
-// &LpExpr - (LpExpr, &LpExpr, i32)
+// &LpExpr - (LpExpr, &LpExpr, f32)
 impl<'a, T> Sub<T> for &'a LpExpression where T: Into<LpExpression> + Clone {
     type Output = LpExpression;
     fn sub(self, _rhs: T) -> LpExpression {
@@ -76,8 +76,8 @@ impl<'a, T> Sub<T> for &'a LpExpression where T: Into<LpExpression> + Clone {
     }
 }
 
-// i32 - &LpExpr
-impl<'a> Sub<&'a LpExpression> for i32 {
+// f32 - &LpExpr
+impl<'a> Sub<&'a LpExpression> for f32 {
     type Output = LpExpression;
     fn sub(self, _rhs: &'a LpExpression) -> LpExpression {
         SubExpr(Rc::new(LitVal(self)), Rc::new(_rhs.clone()))
@@ -87,22 +87,22 @@ impl<'a> Sub<&'a LpExpression> for i32 {
 impl<'a> Neg for &'a LpExpression {
     type Output = LpExpression;
     fn neg(self) -> LpExpression {
-        MulExpr(Rc::new(LitVal(-1)), Rc::new(self.clone()))
+        MulExpr(Rc::new(LitVal(-1.0)), Rc::new(self.clone()))
     }
 }
 
 
 
-// i32 * LpExpr
-impl Mul<LpExpression> for i32 {
+// f32 * LpExpr
+impl Mul<LpExpression> for f32 {
     type Output = LpExpression;
     fn mul(self, _rhs: LpExpression) -> LpExpression {
         LpExpression::MulExpr(Rc::new(LitVal(self)), Rc::new(_rhs))
     }
 }
 
-// i32 * &LpExp
-impl<'a> Mul<&'a LpExpression> for i32 {
+// f32 * &LpExp
+impl<'a> Mul<&'a LpExpression> for f32 {
     type Output = LpExpression;
 
     fn mul(self, _rhs: &'a LpExpression) -> LpExpression {
